+++ coordinator/statement_executor_test.go
@@ -319,12 +319,14 @@ type TSDBStore struct {
 	RestoreShardFn func(id uint64, r io.Reader) error
 	BackupShardFn  func(id uint64, since time.Time, w io.Writer) error
 
-	DeleteDatabaseFn        func(name string) error
-	DeleteMeasurementFn     func(database, name string) error
-	DeleteRetentionPolicyFn func(database, name string) error
-	DeleteShardFn           func(id uint64) error
-	DeleteSeriesFn          func(database string, sources []influxql.Source, condition influxql.Expr) error
-	ShardGroupFn            func(ids []uint64) tsdb.ShardGroup
+	DeleteDatabaseFn          func(name string) error
+	DeleteMeasurementFn       func(database, name string) error
+	DeleteRetentionPolicyFn   func(database, name string) error
+	DeleteShardFn             func(id uint64) error
+	DeleteSeriesFn            func(database string, sources []influxql.Source, condition influxql.Expr) error
+	ShardGroupFn              func(ids []uint64) tsdb.ShardGroup
+	MeasurementsCardinalityFn func(database string) (int64, error)
+	SeriesCardinalityFn       func(database string) (int64, error)
 }
 
 func (s *TSDBStore) CreateShard(database, policy string, shardID uint64, enabled bool) error {
@@ -378,8 +380,16 @@ func (s *TSDBStore) MeasurementNames(database string, cond influxql.Expr) ([][]b
 	return nil, nil
 }
 
+func (s *TSDBStore) MeasurementsCardinality(database string) (int64, error) {
+	return s.MeasurementsCardinalityFn(database)
+}
+
 func (s *TSDBStore) TagValues(_ query.Authorizer, database string, cond influxql.Expr) ([]tsdb.TagValues, error) {
 	return nil, nil
+}
+
+func (s *TSDBStore) SeriesCardinality(database string) (int64, error) {
+	return s.SeriesCardinalityFn(database)
 }
 
 type MockShard struct {