+++ services/retention/service.go
@@ -22,9 +22,9 @@ type Service struct {
 		DeleteShard(shardID uint64) error
 	}
 
-	checkInterval time.Duration
-	wg            sync.WaitGroup
-	done          chan struct{}
+	config Config
+	wg     sync.WaitGroup
+	done   chan struct{}
 
 	logger zap.Logger
 }
@@ -32,26 +32,36 @@ type Service struct {
 // NewService returns a configured retention policy enforcement service.
 func NewService(c Config) *Service {
 	return &Service{
-		checkInterval: time.Duration(c.CheckInterval),
-		done:          make(chan struct{}),
-		logger:        zap.New(zap.NullEncoder()),
+		config: c,
+		logger: zap.New(zap.NullEncoder()),
 	}
 }
 
 // Open starts retention policy enforcement.
 func (s *Service) Open() error {
-	s.logger.Info(fmt.Sprint("Starting retention policy enforcement service with check interval of ", s.checkInterval))
-	s.wg.Add(2)
-	go s.deleteShardGroups()
-	go s.deleteShards()
+	if !s.config.Enabled || s.done != nil {
+		return nil
+	}
+
+	s.logger.Info(fmt.Sprint("Starting retention policy enforcement service with check interval of ", s.config.CheckInterval))
+	s.done = make(chan struct{})
+
+	s.wg.Add(1)
+	go func() { defer s.wg.Done(); s.run() }()
 	return nil
 }
 
 // Close stops retention policy enforcement.
 func (s *Service) Close() error {
-	s.logger.Info("retention policy enforcement terminating")
+	if !s.config.Enabled || s.done == nil {
+		return nil
+	}
+
+	s.logger.Info("Retention policy enforcement service closing.")
 	close(s.done)
+
 	s.wg.Wait()
+	s.done = nil
 	return nil
 }
 
@@ -60,10 +70,8 @@ func (s *Service) WithLogger(log zap.Logger) {
 	s.logger = log.With(zap.String("service", "retention"))
 }
 
-func (s *Service) deleteShardGroups() {
-	defer s.wg.Done()
-
-	ticker := time.NewTicker(s.checkInterval)
+func (s *Service) run() {
+	ticker := time.NewTicker(time.Duration(s.config.CheckInterval))
 	defer ticker.Stop()
 	for {
 		select {
@@ -71,46 +79,26 @@ func (s *Service) deleteShardGroups() {
 			return
 
 		case <-ticker.C:
-			dbs := s.MetaClient.Databases()
-			for _, d := range dbs {
-				for _, r := range d.RetentionPolicies {
-					for _, g := range r.ExpiredShardGroups(time.Now().UTC()) {
-						if err := s.MetaClient.DeleteShardGroup(d.Name, r.Name, g.ID); err != nil {
-							s.logger.Info(fmt.Sprintf("failed to delete shard group %d from database %s, retention policy %s: %s",
-								g.ID, d.Name, r.Name, err.Error()))
-						} else {
-							s.logger.Info(fmt.Sprintf("deleted shard group %d from database %s, retention policy %s",
-								g.ID, d.Name, r.Name))
-						}
-					}
-				}
-			}
-		}
-	}
-}
-
-func (s *Service) deleteShards() {
-	defer s.wg.Done()
-
-	ticker := time.NewTicker(s.checkInterval)
-	defer ticker.Stop()
-	for {
-		select {
-		case <-s.done:
-			return
-
-		case <-ticker.C:
-			s.logger.Info("retention policy shard deletion check commencing")
+			s.logger.Info("Retention policy shard deletion check commencing.")
 
 			type deletionInfo struct {
 				db string
 				rp string
 			}
 			deletedShardIDs := make(map[uint64]deletionInfo, 0)
+
 			dbs := s.MetaClient.Databases()
 			for _, d := range dbs {
 				for _, r := range d.RetentionPolicies {
-					for _, g := range r.DeletedShardGroups() {
+					for _, g := range r.ExpiredShardGroups(time.Now().UTC()) {
+						if err := s.MetaClient.DeleteShardGroup(d.Name, r.Name, g.ID); err != nil {
+							s.logger.Info(fmt.Sprintf("Failed to delete shard group %d from database %s, retention policy %s: %v. Retry in %v.", g.ID, d.Name, r.Name, err, s.config.CheckInterval))
+							continue
+						}
+
+						s.logger.Info(fmt.Sprintf("Deleted shard group %d from database %s, retention policy %s.", g.ID, d.Name, r.Name))
+
+						// Store all the shard IDs that may possibly need to be removed locally.
 						for _, sh := range g.Shards {
 							deletedShardIDs[sh.ID] = deletionInfo{db: d.Name, rp: r.Name}
 						}
@@ -118,19 +106,19 @@ func (s *Service) deleteShards() {
 				}
 			}
 
+			// Remove shards if we store them locally
 			for _, id := range s.TSDBStore.ShardIDs() {
 				if info, ok := deletedShardIDs[id]; ok {
 					if err := s.TSDBStore.DeleteShard(id); err != nil {
-						s.logger.Error(fmt.Sprintf("failed to delete shard ID %d from database %s, retention policy %s: %s",
-							id, info.db, info.rp, err.Error()))
+						s.logger.Error(fmt.Sprintf("Failed to delete shard ID %d from database %s, retention policy %s: %v. Will retry in %v", id, info.db, info.rp, err, s.config.CheckInterval))
 						continue
 					}
-					s.logger.Info(fmt.Sprintf("shard ID %d from database %s, retention policy %s, deleted",
-						id, info.db, info.rp))
+					s.logger.Info(fmt.Sprintf("Shard ID %d from database %s, retention policy %s, deleted.", id, info.db, info.rp))
 				}
 			}
+
 			if err := s.MetaClient.PruneShardGroups(); err != nil {
-				s.logger.Info(fmt.Sprintf("error pruning shard groups: %s", err))
+				s.logger.Info(fmt.Sprintf("Problem pruning shard groups: %s. Will retry in %v", err, s.config.CheckInterval))
 			}
 		}
 	}